automation/services/mina-bp-stats/leaderboard-bot/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.10
packaging==20.9
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
numpy==1.19.5
pandas==1.1.5
sqlalchemy==1.3.23
psycopg2==2.8.6